🎨 Merge codes into '.hpp' file.
